+++ apps/games/app/(play)/@controls/default.tsx
@@ -1,0 +1,2 @@
+import Controls from "./page"
+export default Controls