+++ apps/games/app/(play)/@controls/games/[code]/[year]/[month]/[day]/page.tsx
@@ -1,0 +1,4 @@
+const Controls = ({ params: _params }: { params: { code: string } }) => {
+    return <div>ⓘ</div>
+}
+export default Controls