+++ apps/games/app/(play)/@title/games/[code]/[year]/[month]/[day]/page.tsx
@@ -1,0 +1,5 @@
+import { GAMES } from "~/games/GAMES"
+const Title = ({ params }: { params: { code: string } }) => {
+    return <>{GAMES[params.code]?.title}</>
+}
+export default Title