+++ apps/games/app/(play)/@title/games/[code]/page.tsx
@@ -1,5 +1,10 @@
-import { GAMES } from "~/games/GAMES"
-const Title = ({ params }: { params: { code: string } }) => {
-    return <>{GAMES[params.code]?.title}</>
+import Link from "next/link"
+import SiteTitle from "~/components/SiteTitle"
+const Title = () => {
+    return (
+        <Link href="/">
+            <SiteTitle /> | Games
+        </Link>
+    )
 }
 export default Title