+++ apps/games/app/(play)/games/[code]/AttributionFallback.tsx
@@ -1,0 +1,3 @@
+export const AttributionFallback = () => {
+    return <h4>Edited by &hellip;</h4>
+}