+++ apps/games/games/GAMES.ts
@@ -6,14 +6,20 @@ export type Game = Readonly<{
     summary: string
     title: string
 }>
+const MIKE_KEESEY = {
+    href: "https://www.phylopic.org/contributors/060f03a9-fafd-4d08-81d1-b8f82080573f/t-michael-keesey-silhouettes",
+    name: "Mike Keesey",
+}
 export const GAMES: Record<string, Game> = {
+    ancestor: {
+        author: MIKE_KEESEY,
+        summary: "Figure out which silhouette represents the ancestor.",
+        title: "Ancestor",
+    },
     "four-clades": {
-        author: {
-            href: "https://www.phylopic.org/contributors/060f03a9-fafd-4d08-81d1-b8f82080573f/t-michael-keesey-silhouettes",
-            name: "Mike Keesey",
-        },
+        author: MIKE_KEESEY,
         summary:
-            'Sort these silhouettes into four clades. (A clade is an ancestor plus all descendants, like "mammals" or "falcons" or "flowering plants".',
+            'Sort sixteen silhouette images into four clades. (A clade is an ancestor plus all descendants, like "mammals" or "flowering plants".)',
         title: "Four Clades",
     },
 }