+++ apps/games/next.config.js
@@ -4,5 +4,4 @@ module.exports = {
         domains: ["images.phylopic.org"],
     },
     reactStrictMode: true,
-    swcMinify: true,
 }