+++ packages/utils-aws/src/s3/objectExists.ts
@@ -1,11 +1,14 @@
 import { HeadObjectCommand, HeadObjectCommandInput, S3Client } from "@aws-sdk/client-s3"
+import { isAWSError } from "../errors"
 export const objectExists = async (client: S3Client, input: HeadObjectCommandInput) => {
     const command = new HeadObjectCommand(input)
     try {
         await client.send(command)
     } catch (e) {
-        // :TODO: Check if 404
-        return false
+        if (isAWSError(e) && e.$metadata.httpStatusCode >= 400 && e.$metadata.httpStatusCode < 500) {
+            return false
+        }
+        throw e
     }
     return true
 }